# specs/openapi.yaml
openapi: 3.1.0
info:
  title: bracket-creator API
  description: |
    Web API for generating Kendo tournament brackets.

    This spec covers two servers:
    - **Bracket Generator** (`serve` command): Stateless, form-based bracket generation.
    - **Mobile/Admin API** (`mobile-app` command): Stateful REST API for live tournament management.

    The Mobile/Admin API routes under `/api/` (excluding `/api/viewer/`) require the
    `X-Tournament-Password` header for authentication.
  version: 2.1.0
servers:
  - url: http://localhost:8080
    description: Bracket Generator (serve command)
  - url: http://localhost:8081
    description: Mobile/Admin API (mobile-app command)
...
paths:
...
  /api/tournament:
    get:
      tags: [tournament]
      summary: Get tournament configuration
      security:
        - TournamentPassword: []
      responses:
        '200':
          description: Tournament configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [tournament]
      summary: Create tournament configuration
      description: |
        Creates a new tournament. Allowed without the `X-Tournament-Password` header when no
        tournament has been configured yet (bootstrap). Once a tournament with a password exists,
        auth is required.

        Validation: `name` is trimmed and must be non-empty after trim (`tournament name is
        required`). `password` must be non-empty (`tournament password is required`) — empty
        Password on disk would let AuthMiddleware's `password != t.Password` check vacuously
        pass for any client sending an empty header. `venue` and `date` are also trimmed but
        may be empty.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tournament'
      responses:
        '201':
          description: Tournament created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          $ref: '#/components/responses/BadRequest'
    put:
      tags: [tournament]
      summary: Update tournament configuration
      description: |
        Updates tournament config. Also allowed without auth during bootstrap (same
        condition as POST): `AuthMiddleware` skips the password check only when
        `tournament.md` is missing OR when the on-disk record is the literal default
        placeholder (`Name == "New Tournament" && Password == ""`). A real-named
        record with an empty password is FAIL-CLOSED — rejected with 403
        (`tournament misconfigured: password is not set`) BEFORE this handler runs;
        operator must repair `tournament.md` out-of-band to recover. See
        `internal/mobileapp/middleware.go` AuthMiddleware for the exact branches.

        Validation matches POST for `name` (rejected if empty after trim).

        **Password preserve semantics**: when the request body omits `password` or sends an
        empty string, the server preserves the currently-stored password. This matches the
        admin UI's `password: pass || undefined` pattern ("leave blank to keep current").
        If after the preserve step the password is STILL empty (operator manually edited
        `tournament.md`, or a fresh PUT against a never-initialized tournament), the
        request is rejected with 400 (`tournament password is required`) to force the
        admin to set a password rather than silently allow the empty-stored-password
        vacuous-pass scenario.
      security:
        - TournamentPassword: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tournament'
      responses:
        '200':
          description: Tournament updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/competitions:
    get:
      tags: [competitions]
      summary: List competitions
      description: Returns an empty array `[]` when no competitions exist (never `null`).
      security:
        - TournamentPassword: []
      responses:
        '200':
          description: Array of competitions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Competition'
    post:
      tags: [competitions]
      summary: Create competition
      description: |
        Creates a new competition. The `id` field is optional — when omitted the server
        derives a URL-friendly slug from `name` (lowercase, spaces to hyphens, max 64 chars).

        Validation (returns 400 with explanatory message):
        - `name` is trimmed and must be non-empty (`competition name is required`).
        - `name` must be unique (case-insensitive) across competitions
          (`competition name "X" already exists`).
        - When `id` is omitted AND `name` slugs to empty (no alphanumerics)
          → 400 (`competition ID is required (could not derive one from name)`).
        - Provided `id` must match the path-parameter pattern (see CompetitionId).

        String fields are trimmed server-side: `name`, `numberPrefix`, `kind`,
        `format`, `poolSizeMode`, `startTime`, `date`.
      security:
        - TournamentPassword: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Competition'
      responses:
        '201':
          description: Competition created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/competitions/{id}:
    parameters:
      - $ref: '#/components/parameters/CompetitionId'
    get:
      tags: [competitions]
      summary: Get competition
      security:
        - TournamentPassword: []
      responses:
        '200':
          description: Competition details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [competitions]
      summary: Update competition (settings OR roster, mutually exclusive)
      description: |
        Behaviour depends on whether the request body carries the
        `players` field:

        **Settings-update (body OMITS `players`):**
        The server applies these fields from the body onto the
        on-disk record: `name`, `date`, `startTime`, `poolSize`,
        `poolWinners`, `poolSizeMode`, `courts`, `roundRobin`,
        `withZekkenName`, `teamSize`, `numberPrefix`, `format`,
        `kind`, `mirror`. String fields are trimmed server-side.
        Rejected with 400 if `name` is empty after trim, if `name`
        collides (case-insensitive) with another competition's name,
        or if `date` is non-empty and not in canonical `DD-MM-YYYY`
        format.

        **Roster-save (body INCLUDES `players`, even an empty array):**
        The server writes `participants.csv` (and clears or extracts
        `seeds.csv` from each player's `seed` field) and IGNORES all
        settings fields in the body. This is how the admin
        participants page saves the roster — its body shape is
        `{ ...c, players: np }` where `c` is a possibly-stale
        frontend snapshot of the competition. Treating settings
        fields as authoritative in this case would silently revert
        any concurrent settings change. To update settings AND the
        roster in one logical operation, issue two PUTs (settings
        first, then roster). Players with `seed: 0` produce no seed
        entry; sending `players: []` clears the roster.

        Returns 404 when the competition does not exist (this endpoint
        never creates).

        Server-managed fields — `status`, `hasParticipantIDs` — are
        IGNORED in the body. Use the dedicated transition endpoints
        (`/start`, `/complete`, `/invalidate`) to change status.
      security:
        - TournamentPassword: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Competition'
      responses:
        '200':
          description: Competition updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [competitions]
      summary: Delete competition
      description: |
        Idempotent — returns 204 even when the competition does not exist.
        Allowed in any non-in-progress state (`setup`, `completed`,
        `invalid`, or unset). In-progress competitions (`pools`/`playoffs`)
        must first be invalidated via `POST /api/competitions/{id}/invalidate`.
      security:
        - TournamentPassword: []
      responses:
        '204':
          description: Competition deleted (or did not exist).
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Competition is in progress and must be invalidated first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/competitions/{id}/schedule:
    put:
      tags: [competitions]
      summary: Save schedule
      security:
        - TournamentPassword: []
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ScheduleEntry'
      responses:
        '200':
          description: Schedule saved.
...
  /api/viewer/tournament:
    get:
      tags: [viewer]
      summary: Get public tournament info
      description: Returns tournament details with password field redacted.
      responses:
        '200':
          description: Public tournament info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/viewer/competitions/{id}:
    get:
      tags: [viewer]
      summary: Get full competition state
      description: Returns config, pools, pool matches, standings, bracket, and schedule.
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
      responses:
        '200':
          description: Full competition state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    TournamentPassword:
      type: apiKey
      in: header
      name: X-Tournament-Password
      description: |
        Tournament admin password. Required for all `/api/` routes except
        `/api/viewer/`, `/api/events`, and the bootstrap PUT/POST
        `/api/tournament` while no tournament has been configured yet.

        Header value must exactly equal the stored tournament password
        (exact-string match; no trimming applied to either side).

        Defense-in-depth: if the stored tournament password is somehow
        empty (e.g. operator manually edited `tournament.md` to land
        an empty `password:` field), AuthMiddleware returns 403
        (`tournament misconfigured: password is not set`) for ALL
        requests rather than letting the `header != stored` check
        vacuously pass for empty header.

        Recovery from this state requires **out-of-band repair**: the
        bootstrap PUT/POST `/api/tournament` exception only applies
        when the loaded record is `nil` or matches the literal
        `Name == "New Tournament" && Password == ""` default — a
        real-named record with empty Password is locked out by the 403
        fail-closed branch. Operator must stop the server, edit
        `tournament.md` to either delete it (forces fresh bootstrap)
        or set a non-empty `password:` field in the YAML front-matter,
        then restart.
  parameters:
    CompetitionId:
      in: path
      name: id
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9][a-zA-Z0-9_-]*$'
        maxLength: 64
      description: |
        Competition slug/ID (e.g. "mens-individual"). Must start with an
        alphanumeric character, contain only alphanumerics / hyphens /
        underscores, and be at most 64 chars. Validated server-side via
        `requireValidCompID`; invalid IDs return 400
        (`competition ID contains invalid characters` / `competition ID too long` /
        `competition ID cannot be empty`). The validation is a path-traversal
        defense: without it, an `id` like `../../../etc/passwd` would resolve
        through `filepath.Clean(filepath.Join(...))` and escape the data
        directory.
...
  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Tournament:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
          description: Tournament name. Trimmed server-side; rejected with 400 if empty after trim or longer than 200 characters.
        date:
          type: string
          maxLength: 10
          description: >
            Event date in DD-MM-YYYY format. Trimmed server-side; rejected
            with 400 if non-empty and not a syntactically valid day in
            that format (Feb 31 etc. → 400) OR if the year falls outside
            1900–2100. The year-range matches the frontend validator
            (web-mobile/js/admin_helpers.jsx MIN_YEAR/MAX_YEAR) so a
            value the API accepts is also one the admin UI can edit. The
            web frontend converts the HTML `<input type="date">` ISO
            output to DMY at the input boundary; direct API clients
            send DMY directly.
        venue:
          type: string
          maxLength: 200
          description: Venue name. Trimmed server-side.
        courts:
          type: array
          items:
            type: string
          description: >
            Court labels (e.g. ["A", "B"]). Each label is a single
            character; 26-court cap (A–Z). Duplicate labels are rejected
            with 400 — the frontend uses labels as identity keys (per-court
            rendering, schedule `byCourt` bucketing, filter values) and
            duplicates would collapse data structures.
        password:
          type: string
          maxLength: 256
          description: >
            Admin password. Required (non-empty) on POST /tournament — an
            empty Password on disk would let AuthMiddleware's
            `password != t.Password` check vacuously pass for any client
            sending an empty `X-Tournament-Password` header.

            On PUT /tournament: omitting the field or sending an empty
            string preserves the currently-stored password ("leave blank
            to keep current"). Sending a non-empty string updates it.

            Omitted from viewer responses (`/api/viewer/tournament`).
        openingBlock:
          type: string
          maxLength: 16
          description: Opening ceremony block label/duration.
        lunchBlock:
          type: string
          maxLength: 16
          description: Lunch break block label/duration.
        closingBlock:
          type: string
          maxLength: 16
          description: Closing ceremony block label/duration.
        clockToElapsedMultiplier:
          type: number
          description: Multiplier converting on-clock match minutes to wall-clock elapsed minutes.
        slowestCourtBufferPct:
          type: integer
          description: Buffer percentage applied to the slowest court when estimating overall duration.
    Competition:
      type: object
      properties:
        id:
          type: string
          maxLength: 64
          description: >
            Unique competition identifier. Must start with an alphanumeric
            character, contain only alphanumerics / hyphens / underscores,
            and be at most 64 chars (same rule as the `:id` path parameter
            — see CompetitionId).
            Optional on POST — the server derives a slug from `name` when omitted.
        name:
          type: string
          maxLength: 200
          description: >
            Competition name. Trimmed server-side and rejected with 400
            (`competition name is required`) if empty after trim or
            longer than 200 characters.
            Uniqueness check (case-insensitive) on POST/PUT — duplicate
            names return 400 (`competition name "X" already exists`).
        kind:
          type: string
          description: "individual or team"
          enum: [individual, team]
        format:
          type: string
          description: "pools or playoffs"
          enum: [pools, playoffs]
        teamSize:
          type: integer
          description: Number of members per team (0 for individual).
        poolSize:
          type: integer
        poolSizeMode:
          type: string
          enum: [max, exact]
        poolWinners:
          type: integer
        roundRobin:
          type: boolean
        courts:
          type: array
          items:
            type: string
          description: >
            Per-competition court labels (subset of Tournament.courts).
            Each label is a single character; 26-court cap; duplicates
            rejected with 400. Empty array means "use the tournament
            default" — the engine defaults a competition with no Courts
            to 1 court.
        startTime:
          type: string
          maxLength: 8
          description: First-match start time (HH:MM). Trimmed server-side.
        date:
          type: string
          maxLength: 10
          description: >
            Competition date in DD-MM-YYYY format. Same shape as
            `Tournament.date`: trimmed server-side; rejected with 400 if
            non-empty and not a syntactically valid day in that format
            (Feb 31 etc. → 400) OR if the year falls outside 1900–2100.
        status:
          type: string
          description: >
            Lifecycle state. Set by the server — do not supply on create/update.
            `setup` → not yet started; `pools` → pools phase active;
            `playoffs` → knockout phase active; `completed` → finished;
            `invalid` → operator-declared invalid, prerequisite for deleting
            an already-started competition.
          enum: [setup, pools, playoffs, completed, invalid, ""]
        mirror:
          type: boolean
        withZekkenName:
          type: boolean
        numberPrefix:
          type: string
        hasParticipantIDs:
          type: boolean
        teamMatchType:
          type: string
          description: Team-match scoring model — `fixed` (positions) or `kachinuki` (winner-stays).
          enum: [fixed, kachinuki]
        maxEnchoPeriods:
          type: integer
          description: Maximum number of encho (extended time) periods permitted per match.
        players:
          type: array
          description: >
            Player list. Populated only in viewer responses (`GET /api/viewer/competitions` and
            `GET /api/viewer/competitions/{id}`). Always empty in admin competition responses.
          items:
            $ref: '#/components/schemas/Player'
    Player:
      type: object
      properties:
        id:
          type: string
          maxLength: 64
        name:
          type: string
          maxLength: 100
        displayName:
          type: string
          maxLength: 50
          description: Display name shown on the bracket and Zekken. Defaults to a sanitized form of `name` when omitted.
        dojo:
          type: string
          maxLength: 100
        metadata:
          type: array
          maxItems: 16
          items:
            type: string
            maxLength: 200
        tag:
          type: string
          maxLength: 200
        seed:
          type: integer
          description: Seed rank (0 = unseeded).
        number:
          type: string
          description: Match number prefix + sequence (e.g. "K1"). Only present when number-prefix is set.
...
    Pool:
      type: object
      properties:
        poolName:
          type: string
        players:
          type: array
          items:
            $ref: '#/components/schemas/Player'
    MatchResult:
      type: object
      properties:
        id:
          type: string
        sideA:
          type: string
          maxLength: 100
        sideB:
          type: string
          maxLength: 100
        winner:
          type: string
          maxLength: 100
        ipponsA:
          type: array
          items:
            type: string
          description: "Ippon types scored by side A (M, K, D, T, H)."
        ipponsB:
          type: array
          items:
            type: string
        hansokuA:
          type: integer
        hansokuB:
          type: integer
        decision:
          type: string
          description: >
            Decision code for the match outcome. Extended set:
            `none`/`""` — no decision recorded; `fought` — resolved by scoring;
            `hikiwake` — draw/tie; `kiken` — withdrawal; `fusenpai` — forfeit loss;
            `fusensho` — forfeit win; `daihyosen` — resolved by representative bout;
            `kachinuki-exhaustion` — kachinuki team match ended with one side exhausted.
          enum: [none, fought, hikiwake, kiken, fusenpai, fusensho, daihyosen, kachinuki-exhaustion]
        decisionBy:
          type: string
          description: Which side the decision favours (shiro = white = left, aka = red = right).
          enum: [shiro, aka]
        decisionReason:
          type: string
          maxLength: 200
          description: Free-text operator note for the decision.
        encho:
          $ref: '#/components/schemas/EnchoMetadata'
        status:
          type: string
          enum: [scheduled, running, completed]
        court:
          type: string
          maxLength: 32
        scheduledAt:
          type: string
          maxLength: 32
        subResults:
          type: array
          items:
            $ref: '#/components/schemas/SubMatchResult'
    SubMatchResult:
      type: object
      description: Individual member match result within a team match.
      properties:
        position:
          type: integer
        sideA:
          type: string
          maxLength: 100
        sideB:
          type: string
          maxLength: 100
        ipponsA:
          type: array
          items:
            type: string
        ipponsB:
          type: array
          items:
            type: string
        hansokuA:
          type: integer
        hansokuB:
          type: integer
        winner:
          type: string
          maxLength: 100
        decision:
          type: string
          description: >
            Decision code for this individual sub-match. Same values as `MatchResult.decision`:
            `""` (normal victory) or `"hikiwake"` (draw/tie).
    BracketMatch:
      type: object
      properties:
        id:
          type: string
        sideA:
          type: string
          maxLength: 100
        sideB:
          type: string
          maxLength: 100
        winner:
          type: string
          maxLength: 100
        status:
          type: string
          enum: [scheduled, running, completed]
        court:
          type: string
          maxLength: 32
        scheduledAt:
          type: string
          maxLength: 32
        scoreA:
          type: string
        scoreB:
          type: string
        isOverridden:
          type: boolean
    Bracket:
      type: object
      properties:
        rounds:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/BracketMatch'
    PlayerStanding:
      type: object
      properties:
        player:
          $ref: '#/components/schemas/Player'
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        ipponsGiven:
          type: integer
        ipponsTaken:
          type: integer
        points:
          type: integer
          description: >
            Internal weighted composite sort key — not intended for display.
            Individual formula (×100 for integer arithmetic):
            W×100M − L×1M + D×10K + ipponsGiven×100 − ipponsTaken.
            Team formula:
            W×100B − L×1B + D×10M + IV×100K − IL×10K + IT×1K + PW×100 − PL.
            Use `scoreSummary` for a human-readable breakdown.
        scoreSummary:
          type: string
          description: >
            Human-readable ranking breakdown. Individual format: "W:3 L:1 D:0 | P:5-2".
            Team format: "W:2 L:0 D:1 | IV:8 IL:3 IT:1 | PW:5 PL:3".
        rank:
          type: integer
        isOverridden:
          type: boolean
        individualWins:
          type: integer
        individualLosses:
          type: integer
        individualDraws:
          type: integer
        pointsWon:
          type: integer
        pointsLost:
          type: integer
    ScheduleEntry:
      type: object
      properties:
        matchType:
          type: string
          enum: [pool, bracket, break]
        matchRef:
          type: string
          description: >
            Match identifier (empty for breaks). This value corresponds to the `{mid}` path
            parameter used in match scoring and update endpoints (e.g. PUT /api/competitions/{id}/matches/{mid}/score).
        court:
          type: string
        date:
          type: string
          description: DD-MM-YYYY
        scheduledAt:
          type: string
          description: HH:MM
        status:
          type: string
        isBreak:
          type: boolean
        label:
          type: string
          description: Display label for break entries.
    CompetitionDetail:
      type: object
      description: Full competition state returned by the viewer endpoint.
      properties:
        config:
          $ref: '#/components/schemas/Competition'
        pools:
          type: array
          items:
            $ref: '#/components/schemas/Pool'
        poolMatches:
          type: array
          items:
            $ref: '#/components/schemas/MatchResult'
        standings:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/PlayerStanding'
          description: Map of pool name to ordered standings.
        bracket:
          $ref: '#/components/schemas/Bracket'
        schedule:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleEntry'
        reservedSlots:
          type: array
          items:
            $ref: '#/components/schemas/ReservedSlot'
    ReservedSlot:
      type: object
      properties:
        id:
          type: string
        participantID:
          type: string
        sourceCompID:
          type: string
        sourceRank:
          type: integer
    EnchoMetadata:
      type: object
      description: Metadata about extended-time (encho) periods played for a match.
      properties:
        periodCount:
          type: integer
          minimum: 1
...